// src/types.d.ts
export declare type ShaderIOValue = Vec1 | Vec2 | Vec3 | Vec4 | Vec4[] | Float32Array;
export declare type InstanceIOValue = Vec1 | Vec2 | Vec3 | Vec4;
export declare type UniformIOValue = number | InstanceIOValue | Mat3x3 | Mat4x4 | Float32Array | Texture | number[];
export declare enum InstanceBlockIndex {
    ONE = 1,
    TWO = 2,
    THREE = 3,
    FOUR = 4
}
export declare enum InstanceAttributeSize {
    ONE = 1,
    TWO = 2,
    THREE = 3,
    FOUR = 4,
    ATLAS = 99
}
export declare const instanceAttributeSizeFloatCount: {
    [key: number]: number;
};
export declare enum UniformSize {
    ONE = 1,
    TWO = 2,
    THREE = 3,
    FOUR = 4,
    MATRIX3 = 9,
    MATRIX4 = 16,
    VEC4_ARRAY = 98,
    ATLAS = 99
}
export declare enum VertexAttributeSize {
    ONE = 1,
    TWO = 2,
    THREE = 3,
    FOUR = 4
}
// ...
export interface IVertexAttribute {
    defaults?: number[];
    initWithBuffer?: Float32Array;
    qualifier?: string;
    name: string;
    size: VertexAttributeSize;
    update(vertex: number): ShaderIOValue;
}
export interface IVertexAttributeInternal extends IVertexAttribute {
    materialAttribute: Attribute | null;
}
export interface IInstanceAttribute<T extends Instance> {
    resource?: {
        key(): string;
        name: string;
        shaderInjection?: ShaderInjectionTarget;
    };
    block?: number;
    blockIndex?: InstanceBlockIndex;
    childAttributes?: IInstanceAttribute<T>[];
    easing?: IAutoEasingMethod<Vec>;
    name: string;
    parentAttribute?: IInstanceAttribute<T>;
    qualifier?: string;
    size?: InstanceAttributeSize;
    update(instance: T): InstanceIOValue;
}
// ...
export interface IEasingInstanceAttribute<T extends Instance> extends IInstanceAttribute<T> {
    easing: IAutoEasingMethod<Vec> & {
        uid?: number;
    };
    size: InstanceAttributeSize;
}
export interface IValueInstanceAttribute<T extends Instance> extends IInstanceAttribute<T> {
    atlas: undefined;
}
export declare enum ShaderInjectionTarget {
    VERTEX = 1,
    FRAGMENT = 2,
    ALL = 3
}
export interface IUniform {
    shaderInjection?: ShaderInjectionTarget;
    name: string;
    size: UniformSize;
    qualifier?: string;
    update(uniform: IUniform): UniformIOValue;
}
// ...
export declare type FrameMetrics = {
    currentFrame: number;
    currentTime: number;
    frameDuration: number;
    previousTime: number;
};
export interface IEasingControl {
    readonly delay?: number;
    readonly duration: number;
    readonly end: Vec;
    readonly start: Vec;
    readonly startTime: number;
    setAutomatic(): void;
    setStart(start?: Vec): void;
    setTiming(delay?: number, duration?: number): void;
}
export interface IEasingProps {
    delay?: number;
    duration: number;
    end: Vec;
    isManualStart?: boolean;
    isTimeSet?: boolean;
    start: Vec;
    startTime: number;
}